Move DirHandle, DirectoryEntryReaddir into superblock
